+++ nym-vpn-core/CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - LAN sharing is off by default. Use "Allow LAN" setting to allow it (https://github.com/nymtech/nym-vpn-client/pull/3496)
+- Only select gateways with >= 75% performance, unless the gateway with lower performance is selected explicitly (https://github.com/nymtech/nym-vpn-client/pull/3511)
 
 ### Fixed